Use size_t for option array-based stuff, with (cups_len_t) casts as needed for older versions of CUPS.

+++ pappl/client.h
@@ -21,7 +21,7 @@ extern "C" {
 
 extern char		*papplClientGetCSRFToken(pappl_client_t *client, char *buffer, size_t bufsize) _PAPPL_PUBLIC;
 extern char		*papplClientGetCookie(pappl_client_t *client, const char *name, char *buffer, size_t bufsize) _PAPPL_PUBLIC;
-extern int		papplClientGetForm(pappl_client_t *client, cups_option_t **form) _PAPPL_PUBLIC;
+extern size_t		papplClientGetForm(pappl_client_t *client, cups_option_t **form) _PAPPL_PUBLIC;
 extern const char	*papplClientGetHostName(pappl_client_t *client) _PAPPL_PUBLIC;
 extern int		papplClientGetHostPort(pappl_client_t *client) _PAPPL_PUBLIC;
 extern http_t		*papplClientGetHTTP(pappl_client_t *client) _PAPPL_PUBLIC;
@@ -51,7 +51,7 @@ extern void		papplClientHTMLStartForm(pappl_client_t *client, const char *action
 
 extern http_status_t	papplClientIsAuthorized(pappl_client_t *client) _PAPPL_PUBLIC;
 extern bool		papplClientIsEncrypted(pappl_client_t *client) _PAPPL_PUBLIC;
-extern bool		papplClientIsValidForm(pappl_client_t *client, int num_form, cups_option_t *form) _PAPPL_PUBLIC;
+extern bool		papplClientIsValidForm(pappl_client_t *client, size_t num_form, cups_option_t *form) _PAPPL_PUBLIC;
 
 extern bool		papplClientRespond(pappl_client_t *client, http_status_t code, const char *content_coding, const char *type, time_t last_modified, size_t length) _PAPPL_PUBLIC;
 extern ipp_t		*papplClientRespondIPP(pappl_client_t *client, ipp_status_t status, const char *message, ...) _PAPPL_PUBLIC _PAPPL_FORMAT(3, 4);
